+++ compiler/src/iree/compiler/Dialect/HAL/Utils/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Dialect/HAL/Target/BUILD.bazel                    #
+# compiler/src/iree/compiler/Dialect/HAL/Utils/BUILD.bazel                     #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,25 +12,20 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    Target
+    LLVMLinkerUtils
   HDRS
-    "TargetBackend.h"
-    "TargetDevice.h"
-    "TargetRegistry.h"
+    "LLVMLinkerUtils.h"
   SRCS
-    "TargetBackend.cpp"
-    "TargetDevice.cpp"
-    "TargetRegistry.cpp"
+    "LLVMLinkerUtils.cpp"
   DEPS
+    LLVMBitReader
+    LLVMCore
+    LLVMLinker
     LLVMSupport
-    MLIRIR
-    MLIRPass
-    MLIRSupport
-    MLIRTransforms
-    iree::compiler::Dialect::Flow::IR
+    LLVMTarget
+    LLVMipo
+    iree::compiler::Codegen::Utils
     iree::compiler::Dialect::HAL::IR
-    iree::compiler::Dialect::Util::IR
-    iree::compiler::Utils
   PUBLIC
 )
